NN-5935 refactor punishments further
