+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/controllers/reported/PunishmentsControllerTest.kt
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentMatchers
 import org.mockito.kotlin.any
-import org.mockito.kotlin.anyOrNull
 import org.mockito.kotlin.atLeastOnce
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
@@ -19,14 +18,9 @@ import org.springframework.test.web.servlet.ResultActions
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.TestControllerBase
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.AdditionalDaysDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.PunishmentDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.PunishmentScheduleDto
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.SuspendedPunishmentDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.PunishmentType
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.AdjudicationDomainEventType
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported.PunishmentsService
-import java.time.LocalDate
 
 @WebMvcTest(
   PunishmentsController::class,
@@ -176,170 +170,7 @@ class PunishmentsControllerTest : TestControllerBase() {
     }
   }
 
-  @Nested
-  inner class GetSuspendedPunishments {
-    @BeforeEach
-    fun beforeEach() {
-      whenever(
-        punishmentsService.getSuspendedPunishments(
-          any(),
-          anyOrNull(),
-        ),
-      ).thenReturn(SUSPENDED_PUNISHMENTS_DTO)
-    }
-
-    @Test
-    fun `responds with a unauthorised status code`() {
-      getSuspendedPunishmentsRequest().andExpect(MockMvcResultMatchers.status().isUnauthorized)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["SCOPE_write"])
-    fun `responds with a forbidden status code for non ALO`() {
-      getSuspendedPunishmentsRequest().andExpect(MockMvcResultMatchers.status().isForbidden)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["ROLE_ADJUDICATIONS_REVIEWER"])
-    fun `responds with a forbidden status code for ALO without write scope`() {
-      getSuspendedPunishmentsRequest().andExpect(MockMvcResultMatchers.status().isForbidden)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["ROLE_ADJUDICATIONS_REVIEWER", "SCOPE_write"])
-    fun `makes a call to get a set of suspended punishments`() {
-      getSuspendedPunishmentsRequest()
-        .andExpect(MockMvcResultMatchers.status().isOk)
-
-      verify(punishmentsService).getSuspendedPunishments("AE1234", "12345")
-    }
-
-    private fun getSuspendedPunishmentsRequest(): ResultActions {
-      return mockMvc
-        .perform(
-          MockMvcRequestBuilders.get("/reported-adjudications/punishments/AE1234/suspended/v2?chargeNumber=12345")
-            .header("Content-Type", "application/json"),
-        )
-    }
-  }
-
-  @Nested
-  inner class GetReportsWithAdditionalDays {
-    @BeforeEach
-    fun beforeEach() {
-      whenever(
-        punishmentsService.getReportsWithAdditionalDays(
-          any(),
-          any(),
-          any(),
-        ),
-      ).thenReturn(ADDITIONAL_DAYS_DTO)
-    }
-
-    @Test
-    fun `responds with a unauthorised status code`() {
-      getReportsWithAdditionalDaysRequest().andExpect(MockMvcResultMatchers.status().isUnauthorized)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["SCOPE_write"])
-    fun `responds with a forbidden status code for non ALO`() {
-      getReportsWithAdditionalDaysRequest().andExpect(MockMvcResultMatchers.status().isForbidden)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["ROLE_ADJUDICATIONS_REVIEWER"])
-    fun `responds with a forbidden status code for ALO without write scope`() {
-      getReportsWithAdditionalDaysRequest().andExpect(MockMvcResultMatchers.status().isForbidden)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["ROLE_ADJUDICATIONS_REVIEWER", "SCOPE_write"])
-    fun `makes a call to get reports with additional days`() {
-      getReportsWithAdditionalDaysRequest()
-        .andExpect(MockMvcResultMatchers.status().isOk)
-
-      verify(punishmentsService).getReportsWithAdditionalDays("12345", "AE1234", PunishmentType.ADDITIONAL_DAYS)
-    }
-
-    private fun getReportsWithAdditionalDaysRequest(): ResultActions {
-      return mockMvc
-        .perform(
-          MockMvcRequestBuilders.get("/reported-adjudications/punishments/AE1234/for-consecutive?type=ADDITIONAL_DAYS&chargeNumber=12345")
-            .header("Content-Type", "application/json"),
-        )
-    }
-  }
-
-  @Nested
-  inner class GetActivePunishments {
-    @BeforeEach
-    fun beforeEach() {
-      whenever(
-        punishmentsService.getActivePunishments(
-          ArgumentMatchers.anyLong(),
-        ),
-      ).thenReturn(
-        listOf(
-          ActivePunishmentDto(
-            chargeNumber = "1234",
-            punishmentType = PunishmentType.REMOVAL_WING,
-          ),
-        ),
-      )
-    }
-
-    @Test
-    fun `responds with a unauthorised status code`() {
-      activePunishmentsRequest().andExpect(MockMvcResultMatchers.status().isUnauthorized)
-    }
-
-    @Test
-    @WithMockUser(username = "ITAG_USER", authorities = ["ROLE_VIEW_ADJUDICATIONS"])
-    fun `makes a call to get active punishments`() {
-      activePunishmentsRequest()
-        .andExpect(MockMvcResultMatchers.status().isOk)
-
-      verify(punishmentsService).getActivePunishments(any())
-    }
-
-    private fun activePunishmentsRequest(): ResultActions {
-      return mockMvc
-        .perform(
-          MockMvcRequestBuilders.get("/reported-adjudications/punishments/1234567/active")
-            .header("Content-Type", "application/json"),
-        )
-    }
-  }
-
   companion object {
     val PUNISHMENT_REQUEST = PunishmentRequest(type = PunishmentType.REMOVAL_ACTIVITY, days = 10)
-    val PUNISHMENT_COMMENT_REQUEST = PunishmentCommentRequest(comment = "some text")
-    val SUSPENDED_PUNISHMENTS_DTO = listOf(
-      SuspendedPunishmentDto(
-        chargeNumber = "1",
-        corrupted = false,
-        punishment =
-        PunishmentDto(
-          type = PunishmentType.REMOVAL_WING,
-          schedule = PunishmentScheduleDto(
-            days = 10,
-            suspendedUntil = LocalDate.now(),
-          ),
-        ),
-      ),
-    )
-    val ADDITIONAL_DAYS_DTO = listOf(
-      AdditionalDaysDto(
-        chargeNumber = "1",
-        chargeProvedDate = LocalDate.now(),
-        punishment = PunishmentDto(
-          type = PunishmentType.ADDITIONAL_DAYS,
-          schedule = PunishmentScheduleDto(
-            days = 10,
-          ),
-        ),
-      ),
-    )
   }
 }
